third_party/WebKit/LayoutTests/http/tests/websocket/connect-error-by-no-websocket-server-closeonerror-expected.txt

CONSOLE WARNING: WebSocket connection to 'ws://127.0.0.1:8889/' failed: WebSocket is closed before the connection is established.
Test that WebSocket fires error event if no WebSocket Server can be connected, and connection is closed from onerror event

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Test start (Port 8889)
PASS onerror was called
PASS onclose was called
PASS successfullyParsed is true

TEST COMPLETE
